Split out notifications tests, aligning with source package structure
